Keeping it SMOL
With Smart Methods of Leporellion (SMOL) we want to simplify the Principles and Practices of the Leporellion by using a complete Runicons references and Basic Leporellion charset. SMOL can be handwritten and has its own grammar. You need to download the assets.zip folder and the presented runicons.html to preview the embedded Keynote "SMOL - Complete Runicons Reference and Basic Leporellion Charset" (at the very beginning) in HTML