runicodes vs runicons
Work with SMOL runicodes without using a font. Before using a SMOL runicons you must have a special font type installed or need a deeper understanding of character sets, e.g. prophecy symbols or Smart Message abbrevations. The Keynote presentation of Runicons in HTML in the asset.zip is no longer supported and will be delete soon. Filter and sort directly in an numbers or excel file when you want to learn. The list include up to a 1000 words, phrases including an alphabetical arrangement in NewMorse and dialects (Ideogrammar, Emoticons, Wellish) and so on. Next, I set forth the list with VisualBraille ...