SMOL Font Family starts with style T2Hue
See color space and the most relevant #hexvalue in the file .rtf or .pdf. The whole Latin charset in WINDOWS and MAC are available as a colored OpenTypeFont and so on. To be continued ... what next? The Type 'QuantumScript' coming soon. All styles then replaces the styles like 'normal', 'regular' or 'standard'. Switching between styles and fonts into another translate the text. So SMOL Font Family is a translatable encoding format.