Add new SMOL Style: QuantumScript
QuantumScript is a 8-bit sequence that based on monochrome 1 and 0. Reduce line spacing to show that is similar looking to a quick response (QR). QuantumScript can be seen as an alternative Morse Code. I am using Inkscape to build the glyphs 92 x 92 then import them into BirdFont to create the fonts.